@@ -1,80 +1,90 @@
 {
-  "name": "openwerewolf",
-  "version": "0.0.0",
-  "description": "An open source, online multiplayer werewolf game.",
-  "main": "app.js",
-  "dependencies": {
-    "@types/jquery": "^3.3.29",
-    "@types/node": "^9.6.35",
-    "@types/pixi.js": "^4.8.6",
-    "@types/react": "^16.7.20",
-    "@types/react-dom": "^16.0.11",
-    "@types/semantic-ui": "^2.2.6",
-    "@types/simplebar": "^2.4.1",
-    "@types/socket.io": "^1.4.40",
-    "@types/socket.io-client": "^1.4.32",
-    "@types/webfontloader": "^1.6.29",
-    "bcrypt": "^3.0.2",
-    "connect-redis": "^3.4.0",
-    "csstype": "^2.6.1",
-    "express": "^4.16.4",
-    "express-session": "^1.15.6",
-    "grawlix": "^1.0.6",
-    "mysql": "^2.16.0",
-    "nodemailer": "^4.6.8",
-    "pixi.js": "^4.8.5",
-    "pug": "^2.0.3",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "redis": "^2.8.0",
-    "redis-server": "^1.2.2",
-    "selenium-webdriver": "^3.6.0",
-    "semantic-ui": "^2.4.2",
-    "simplebar": "^3.0.0-beta.5",
-    "socket.io": "^2.1.1",
-    "systemjs": "^3.0.0",
-    "ts-loader": "^5.3.3",
-    "typescript": "^3.2.2",
-    "webfontloader": "^1.6.28",
-    "webpack": "^4.29.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^24.0.11",
-    "jest": "^24.7.1",
-    "mocha": "^5.2.0",
-    "ts-jest": "^24.0.1",
-    "webpack-cli": "^3.3.0"
-  },
-  "scripts": {
-    "test": "jest --silent --verbose",
-    "watch:debug": "tsc \n webpack --watch & \n tsc --watch & \n sass --watch client/css/main.scss:client/css/main.css & \n nodemon app.js debug",
-    "watch": "tsc \n webpack --watch & \n tsc --watch & \n sass --watch client/css/main.scss:client/css/main.css & \n nodemon app.js",
-    "watch:database": "tsc \n webpack --watch & \n tsc --watch & \n sass --watch client/css/main.scss:client/css/main.css & \n nodemon app.js database",
-    "builddocs": "typedoc --out docs --mode 'file' --excludeExternals --externalPattern \"**/node_modules/**\" --ignoreCompilerErrors ../OpenWerewolf/Core --target es6 --theme default --readme none \n touch docs/.nojekyll",
-    "build": "tsc \n webpack \n sass client/css/main.scss client/css/main.css",
-    "start": "echo \"*** Please 'npm run build' first if you haven't already!***\" \n node app.js",
-    "start:database": "echo \"*** Please 'npm run build' first if you haven't already!***\" \n node app.js database",
-    "start:debug": "echo \"*** Please 'npm run build' first if you haven't already!***\" \n node app.js debug"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/JamesCraster/OpenWerewolf.git"
-  },
-  "keywords": [
-    "mafia",
-    "werewolves",
-    "multiplayer",
-    "online",
-    "game",
-    "party",
-    "deception",
-    "werewolf",
-    "detective"
-  ],
-  "author": "James Vaughan Craster",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/JamesCraster/OpenWerewolf/issues"
-  },
-  "homepage": "https://github.com/JamesCraster/OpenWerewolf#readme"
+    "name": "openwerewolf",
+    "version": "0.0.0",
+    "description": "An open source, online multiplayer werewolf game.",
+    "type": "module",
+    "main": "dist/app.js",
+    "dependencies": {
+        "@types/jquery": "^3.3.29",
+        "@types/node": "^20.10.3",
+        "@types/pixi.js": "^4.8.6",
+        "@types/react": "^16.7.20",
+        "@types/react-dom": "^16.0.11",
+        "@types/semantic-ui": "^2.2.6",
+        "@types/simplebar": "^2.4.1",
+        "@types/socket.io": "^3.0.2",
+        "@types/socket.io-client": "^3.0.0",
+        "@types/webfontloader": "^1.6.29",
+        "buffer": "^6.0.3",
+        "connect-redis": "^7.1.0",
+        "crypto-browserify": "^3.12.0",
+        "csstype": "^2.6.1",
+        "express": "^4.21.2",
+        "express-session": "^1.17.3",
+        "grawlix": "^1.0.6",
+        "ioredis": "^5.3.2",
+        "mysql2": "^3.6.5",
+        "nodemailer": "^4.6.8",
+        "path-browserify": "^1.0.1",
+        "pixi.js": "^4.8.5",
+        "process": "^0.11.10",
+        "prop-types": "^15.8.1",
+        "pug": "^2.0.3",
+        "react": "^16.7.0",
+        "react-dom": "^16.7.0",
+        "simplebar": "^3.0.0-beta.5",
+        "socket.io": "^4.7.2",
+        "stream-browserify": "^3.0.0",
+        "systemjs": "^3.0.0",
+        "ts-loader": "^9.5.1",
+        "typescript": "^5.3.2",
+        "url": "^0.11.3",
+        "util": "^0.12.5",
+        "webfontloader": "^1.6.28",
+        "webpack": "^5.89.0"
+    },
+    "devDependencies": {
+        "@types/babel-types": "^7.0.15",
+        "@types/connect-redis": "^0.0.23",
+        "@types/express": "^4.17.21",
+        "@types/express-session": "^1.17.10",
+        "@types/jest": "^29.5.10",
+        "@types/prop-types": "^15.7.11",
+        "jest": "^29.7.0",
+        "mocha": "^10.2.0",
+        "sass": "^1.69.5",
+        "ts-jest": "^29.1.1",
+        "ts-node": "^10.9.2",
+        "ts-node-dev": "^2.0.0",
+        "webpack-cli": "^5.1.4"
+    },
+    "scripts": {
+        "test": "jest --silent --verbose",
+        "dev": "NODE_OPTIONS='--loader ts-node/esm --no-warnings' ts-node-dev --respawn --transpile-only app.ts",
+        "build": "webpack --config webpack.config.cjs --mode production && sass client/css/main.scss client/css/main.css",
+        "start": "node dist/app.js",
+        "watch:sass": "sass --watch client/css/main.scss:client/css/main.css",
+        "watch": "webpack --config webpack.config.cjs --watch --mode development"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/JamesCraster/OpenWerewolf.git"
+    },
+    "keywords": [
+        "mafia",
+        "werewolves",
+        "multiplayer",
+        "online",
+        "game",
+        "party",
+        "deception",
+        "werewolf",
+        "detective"
+    ],
+    "author": "James Vaughan Craster",
+    "license": "Apache-2.0",
+    "bugs": {
+        "url": "https://github.com/JamesCraster/OpenWerewolf/issues"
+    },
+    "homepage": "https://github.com/JamesCraster/OpenWerewolf#readme"
 }
