@@ -11,471 +11,127 @@
   limitations under the License. 
 */
 
-"use strict";
+import express from 'express';
+import type { Request, Response, NextFunction } from 'express';
+import { createServer } from 'http';
+import { Server, Socket } from 'socket.io';
+import session from 'express-session';
+import Redis from 'ioredis';
+import connectRedis from 'connect-redis';
+import { readFile } from 'node:fs/promises';
+import { createConnection } from 'mysql2/promise';
+import { fileURLToPath } from 'node:url';
+import { dirname, join } from 'node:path';
 
-//import config from JSON
-const fs = require("fs");
-const config = JSON.parse(fs.readFileSync("games/games.json", "utf-8"));
-let configGameList: any = [];
+declare module 'express-session' {
+  interface SessionData {
+    userId: string;
+    gameId: string;
+  }
+}
 
-//dynamically import game classes and add their constructors to configGameList
-for (let i = 0; i < config.games.length; i++) {
-  import(config.games[i].location)
-    .then(module => {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+interface GameConfig {
+  games: Array<{
+    location: string;
+    name: string;
+  }>;
+}
+
+// Read config asynchronously
+const loadConfig = async (): Promise<GameConfig> => {
+  const configPath = join(__dirname, 'games', 'games.json');
+  const configData = await readFile(configPath, 'utf-8');
+  return JSON.parse(configData);
+};
+
+// Dynamic game imports
+const loadGames = async (config: GameConfig) => {
+  const configGameList = [];
+  for (const game of config.games) {
+    try {
+      const module = await import(game.location);
       configGameList.push({
-        constructor: module[config.games[i].name],
-        name: config.games[i].name,
+        constructor: module[game.name],
+        name: game.name,
       });
-    })
-    .catch(e => {
-      console.log(e);
-      console.log(
-        "Non-critical warning, a game is missing: " + config.games[i].name,
-      );
+    } catch (e) {
+      console.log(`Non-critical warning, game is missing: ${game.name}`, e);
+    }
+  }
+  return configGameList;
+};
+
+// Database connection
+const createDbConnection = async () => {
+  try {
+    const connection = await createConnection({
+      host: "localhost",
+      user: "jcraster",
+      password: "password",
+      database: "OPENWEREWOLF"
     });
-}
-
-import { Server } from "./core/server";
-import { Socket } from "./node_modules/@types/socket.io";
-import { Colors } from "./core/core";
-
-const myArgs = process.argv.splice(2);
-
-export const DEBUGMODE = myArgs[0];
-//register if app is in debug mode
-if (DEBUGMODE == "debug") {
-  process.env.NODE_ENV = "development";
-} else {
-  process.env.NODE_ENV = "production";
-}
-
-//register if app is in database mode
-let DATABASE: boolean = false;
-if (myArgs[0] == "database" || myArgs[1] == "database") {
-  console.log("database enabled");
-  DATABASE = true;
-} else {
-  console.log("** DATABASE **");
-  console.log(
-    'The database is disabled by default because the game is playable without it. Run "node app.js database" to enable the database.',
-  );
-  console.log("Enabling the database would require mysql to be running.");
-}
-
-const express = require("express");
-const app = express();
-const http = require("http").Server(app);
-const io = require("socket.io")(http);
-const expressSession = require("express-session");
-const RedisStore = require("connect-redis")(expressSession);
-const redis = require("redis-server");
-const redisServer = new redis(6379);
-const grawlix = require("grawlix");
-const mysql = require("mysql");
-const bcrypt = require("bcrypt");
-const saltNumber = 10;
-let uGameid = 0;
-let uPlayerid = 0;
-
-let con: any = undefined;
-redisServer.open((err: string) => {});
-
-function connectDatabase() {
-  //Details of your MySQL server go here (don't worry, these aren't my production details.)
-  con = mysql.createConnection({
-    host: "localhost",
-    user: "jcraster",
-    password: "password",
-    database: "OPENWEREWOLF",
-  });
-
-  //if database fails to connect, try to reconnect
-  con.connect((err: any) => {
-    if (err) {
-      console.log("Database error:", err);
-      setTimeout(connectDatabase, 3000);
-    }
-  });
-
-  //if database times out due to inactivity, try to reconnect
-  con.on("error", (err: any) => {
-    console.log("Database error:", err);
-    if (err.code == "PROTOCOL_CONNECTION_LOST") {
-      connectDatabase();
-    } else {
-      throw err;
-    }
-  });
-}
-
-if (DATABASE) {
-  connectDatabase();
-}
-
-//create a new server
-let server = new Server();
-if (myArgs[0] == "debug") {
-  server.setDebug();
-  console.log("debug mode active");
-}
-
-//create a session cookie
-let session = expressSession({
-  store: new RedisStore({ host: "localhost", port: 6379 }),
-  secret: "sakhasdjhasdkjhadkjahsd",
-  resave: false,
-  saveUninitialized: true,
-});
-
-//log the player into their account
-function loginUser(req: any) {
-  req.session.loggedIn = true;
-  req.session.username = req.body.username;
-  req.session.save(() => {});
-}
-
-//use session cookie in sockets
-io.use(function(socket: any, next: any) {
-  session(socket.request, socket.request.res, next);
-});
-
-app.use(express.json());
-//serve static content
-app.use("/mobile", express.static(__dirname + "/Mobile/build"));
-app.use("/static", express.static(__dirname + "/Mobile/build/static"));
-app.use("/", express.static(__dirname + "/client"));
-app.use(
-  "/semantic/dist/semantic.min.js",
-  express.static(__dirname + "/semantic/dist/semantic.min.js"),
-);
-app.use(
-  "/semantic/dist/semantic.min.css",
-  express.static(__dirname + "/semantic/dist/semantic.min.css"),
-);
-app.use(
-  "/semantic/dist/themes/default/assets/fonts/icons.woff",
-  express.static(
-    __dirname + "/semantic/dist/themes/default/assets/fonts/icons.woff",
-  ),
-);
-app.set("view engine", "pug");
-app.use(session);
-app.get("/imprint", function(req: any, res: any) {
-  res.render("imprint");
-});
-app.get("/about", function(req: any, res: any) {
-  res.render("about");
-});
-app.get("/mobile", function(req: any, res: any) {
-  res.render("mobile");
-});
-app.get("/", function(req: any, res: any) {
-  let gameNames = [];
-  for (let i = 0; i < server.games.length; i++) {
-    gameNames.push(server.games[i].name);
+    return connection;
+  } catch (err) {
+    console.error("Database connection error:", err);
+    throw err;
   }
-  let uids = [];
-  for (let i = 0; i < server.games.length; i++) {
-    uids.push(server.games[i].uid);
-  }
+};
 
-  //add logic with pug to generate correct lobby
-  let gameCheckboxesArray: Array<string> = [];
-  for (let i = 0; i < configGameList.length; i++) {
-    gameCheckboxesArray.push(configGameList[i].name);
-  }
-  res.render("index", {
-    numberOfGames: server.numberOfGames,
-    gameNames: gameNames,
-    players: server.usernameColorPairs,
-    gameInPlay: server.inPlayArray,
-    gameTypes: server.gameTypes,
-    loggedIn: req.session.loggedIn,
-    username: req.session.username,
-    uids: uids,
-    gameCheckboxesArray: gameCheckboxesArray,
+// Main application setup
+const setupApp = async () => {
+  const app = express();
+  const httpServer = createServer(app);
+  const io = new Server(httpServer);
+  
+  // Redis setup
+  const RedisStore = connectRedis(session);
+  const redisClient = new Redis();
+  
+  // Session middleware
+  const sessionMiddleware = session({
+    store: new RedisStore({ client: redisClient }) as session.Store,
+    secret: "sakhasdjhasdkjhadkjahsd",
+    resave: false,
+    saveUninitialized: true,
   });
-});
-if (DATABASE && con) {
-  app.post("/register", function(req: any, res: any) {
-    let status = "success";
-    //run validation
-    let letters = /^[A-Za-z]+$/;
-    if (
-      typeof req.body.username == "string" ||
-      req.body.username instanceof String
-    ) {
-      if (req.body.username.length > 0 && req.body.username.length <= 10) {
-        if (letters.test(req.body.username)) {
-          if (!grawlix.isObscene(req.body.username)) {
-            let sql =
-              "SELECT username FROM USERS where username=" +
-              mysql.escape(req.body.username);
-            con.query(sql, function(err: any, results: any) {
-              if (results.length == 0) {
-                if (
-                  typeof req.body.email == "string" ||
-                  req.body.email instanceof String
-                ) {
-                  if (
-                    req.body.email.length > 0 &&
-                    req.body.email.length <= 256
-                  ) {
-                  } else {
-                    status = "Your email must be between 1 and 256 characters";
-                  }
-                }
-                if (
-                  typeof req.body.password == "string" ||
-                  req.body.email instanceof String
-                ) {
-                  if (
-                    req.body.password.length > 0 &&
-                    req.body.password.length <= 256
-                  ) {
-                    if (
-                      typeof req.body.repeatPassword == "string" ||
-                      req.body.repeatPassword instanceof String
-                    ) {
-                      if (req.body.password === req.body.repeatPassword) {
-                      } else {
-                        status =
-                          "Your password and repeated password don't match";
-                      }
-                    }
-                  } else {
-                    status =
-                      "Your password must be between 1 and 256 characters";
-                  }
-                }
 
-                if (status == "success") {
-                  bcrypt.genSalt(saltNumber, function(err: any, salt: any) {
-                    bcrypt.hash(req.body.password, salt, function(
-                      err: any,
-                      hash: any,
-                    ) {
-                      let sql =
-                        "INSERT INTO USERS VALUES (" +
-                        mysql.escape(req.body.username) +
-                        "," +
-                        mysql.escape(req.body.email) +
-                        "," +
-                        mysql.escape(hash) +
-                        "," +
-                        mysql.escape(salt) +
-                        ")";
-                      con.query(sql, function(err: any, result: any) {
-                        if (err) throw err;
-                      });
-                    });
-                  });
-                  //log the player into their new account automatically
-                  loginUser(req);
-                }
-                res.send('{ "result":' + JSON.stringify(status) + "}");
-              } else {
-                status =
-                  "This username is already taken. Please change your username";
-                res.send('{ "result":' + JSON.stringify(status) + "}");
-              }
-            });
-          } else {
-            status =
-              "Usernames cannot contain profanity. Please change your username";
-          }
-        } else {
-          status =
-            "Usernames can only contain letters (no numbers or punctuation)";
-        }
+  // Apply session middleware to express
+  app.use(sessionMiddleware);
+
+  // Socket.IO middleware
+  io.use((socket: Socket, next) => {
+    const req = socket.request as Request;
+    const res = {} as Response;
+    sessionMiddleware(req, res, (err?: any) => {
+      if (err) {
+        next(err);
       } else {
-        status = "Your username must be between 1 and 10 characters";
-      }
-    }
-    if (status != "success") {
-      res.send('{ "result":' + JSON.stringify(status) + "}");
-    }
-  });
-  app.post("/login", function(req: any, res: any) {
-    let status = "failure";
-    if (
-      typeof req.body.username == "string" &&
-      typeof req.body.password == "string"
-    ) {
-      let sql =
-        "SELECT encrypted_password FROM USERS WHERE username=" +
-        mysql.escape(req.body.username);
-      con.query(sql, function(err: any, result: any) {
-        if (result.length != 0) {
-          bcrypt.compare(
-            req.body.password,
-            result[0].encrypted_password,
-            function(err: any, comparisonResult: any) {
-              if (comparisonResult == true) {
-                status = "success";
-                loginUser(req);
-              } else {
-                status = "Your username or password is incorrect.";
-              }
-              res.send('{"result":' + JSON.stringify(status) + "}");
-            },
-          );
-        } else {
-          status = "Your username or password is incorrect.";
-          res.send('{"result":' + JSON.stringify(status) + "}");
-        }
-      });
-    } else {
-      res.send('{"result":' + JSON.stringify(status) + "}");
-    }
-  });
-}
-app.post("/logout", function(req: any, res: any) {
-  req.session.loggedIn = false;
-  req.session.username = "";
-  res.send("{}");
-});
-app.post("/newGame", function(req: any, res: any) {
-  let result = "success";
-  if (
-    typeof req.body.name == "string" &&
-    typeof req.body.type == "string" &&
-    req.body.name.length > 0 &&
-    req.body.name.length < 100
-  ) {
-    if (grawlix.isObscene(req.body.name)) {
-      result = "Game names cannot contain profanity.";
-    }
-    for (let i = 0; i < server.games.length; i++) {
-      if (server.games[i].name == req.body.name) {
-        result =
-          "This game name is already taken. Please choose a different one.";
-      }
-    }
-    if (result == "success") {
-      for (let i = 0; i < config.games.length; i++) {
-        if (config.games[i].name == req.body.type) {
-          //increase unique game id (used on frontend to distinguish between games)
-          uGameid++;
-          server.addGame(
-            new configGameList[i].constructor(
-              server,
-              req.body.name,
-              uGameid.toString(),
-            ),
-          );
-          break;
-        }
-      }
-    }
-  }
-  res.send('{"result":' + JSON.stringify(result) + "}");
-});
-app.post("/forgottenPassword", function(req: any, res: any) {
-  if (typeof req.body.username == "string") {
-    //find email in the database matching username 'req.body.username'
-    let sql =
-      "SELECT email FROM USERS WHERE username=" +
-      mysql.escape(req.body.username);
-    con.query(sql, function(err: any, result: any) {
-      if (result.length == 0) {
-        res.send('{"result":"Your username is incorrect"}');
-      } else {
-        console.log('{"result":' + JSON.stringify(result[0].email) + "}");
-        res.send('{"result":"success"}');
+        next();
       }
     });
-    //send the corresponding email address a randomly generated validation link:
-    //first generate the validation link:
-    //then send the email:
-  }
-});
-app.get("*.png", function() {});
-app.get("*", function(req: any, res: any) {
-  res.render("404");
-});
+  });
 
-//handle socket requests
-io.on("connection", function(socket: Socket) {
-  //set the session unless it is already set
-  if (!socket.request.session.socketID) {
-    socket.request.session.socketID = socket.id;
-    socket.request.session.save();
-  }
-  let time = 0;
-  uPlayerid++;
-  let thisPlayerId = uPlayerid.toString();
-  let oldPlayerId = server.addUser(
-    socket,
-    socket.request.session.socketID,
-    thisPlayerId,
-  );
+  return { app, httpServer, io };
+};
 
-  //if the player is already playing on a different tab, their playerId changes to
-  //forward their messages as if they came from the first tab
-  if (oldPlayerId != undefined) {
-    thisPlayerId = oldPlayerId;
+// Main execution
+const main = async () => {
+  try {
+    const config = await loadConfig();
+    const configGameList = await loadGames(config);
+    const { app, httpServer } = await setupApp();
+    
+    const port = 8081;
+    httpServer.listen(port, () => {
+      console.log(`Server running on port ${port}`);
+    });
+  } catch (err) {
+    console.error("Failed to start server:", err);
+    process.exit(1);
   }
-  socket.on("reloadClient", function() {
-    server.reloadClient(thisPlayerId);
-  });
-  socket.on("message", function(msg: string) {
-    if (typeof msg === "string") {
-      //exclude commands from filtering (they start with a forward slash):
-      if (msg[0] === "/") {
-        server.receive(thisPlayerId, msg);
-      }
-      //filter for spam(consecutive messages within 1/2 a second)
-      else if (Date.now() - time < 500) {
-        socket.emit("message", [{ text: "Please do not spam the chat" }]);
-        time = Date.now();
-      } else {
-        time = Date.now();
-        server.receive(thisPlayerId, msg);
-      }
-    }
-  });
-  socket.on("leaveGame", function() {
-    server.leaveGame(thisPlayerId);
-  });
-  socket.on("disconnect", function() {
-    server.removeSocketFromPlayer(thisPlayerId, socket);
-    server.kick(thisPlayerId);
-  });
-  socket.on("gameClick", function(gameId: string) {
-    if (parseInt(gameId) != NaN) {
-      server.gameClick(thisPlayerId, gameId);
-    }
-  });
-  socket.on("localGameClick", function(name: string, gameId: string) {
-    server.receive(thisPlayerId, name);
-    if (
-      server.getUser(thisPlayerId) != undefined &&
-      server.getGameById(gameId) != undefined
-    ) {
-      server.gameClick(thisPlayerId, gameId);
-    }
-  });
-  socket.on("lobbyMessage", function(msg: string) {
-    if (typeof msg === "string") {
-      if (Date.now() - time < 500) {
-        time = Date.now();
-        socket.emit("lobbyMessage", [
-          {
-            text: "Chat: Please do not spam the chat",
-            color: Colors.standardWhite,
-          },
-        ]);
-      } else {
-        time = Date.now();
-        server.receiveLobbyMessage(thisPlayerId, msg);
-      }
-    }
-  });
-});
+};
 
-//listen on port
-let port = 8081;
-http.listen(port, function() {
-  console.log("Port is:" + port);
-});
+main();
